requirements.txt
alembic==1.16.5
beautifulsoup4==4.12.2
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.1
dnspython==2.4.2
et_xmlfile==2.0.0
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
future==1.0.0
google-genai
googlesearch-python==1.2.3
greenlet==3.2.4
httplib2==0.31.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
lxml==4.9.3
Mako==1.3.10
MarkupSafe==3.0.2
nltk==3.9.1
numpy==2.3.3
openpyxl==3.1.2
pandas==2.2.2
pdfminer.six==20221105
pdfplumber==0.9.0
Pillow==10.1.0
proto-plus==1.26.1
protobuf==6.32.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyMuPDF==1.23.0
PyMuPDFb==1.23.0
pyparsing==3.2.5
pypdf==3.17.1
PyPDF2==3.0.1
pyphen==0.17.2
pypng==0.20220715.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-pptx==0.6.21
python-whois==0.8.0
pytrends==4.9.2
pytz==2025.2
qrcode==7.4.2
regex==2025.9.18
reportlab==4.0.4
requests==2.31.0
rsa==4.9.1
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.43
textstat==0.7.3
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
Wand==0.6.13
Werkzeug==3.0.1
WTForms==3.1.1
xlsxwriter==3.2.9